layers/l2_brain/pyproject.toml
[project]
name = "brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kuzu>=0.11.3",
    "maturin>=1.13.1",
    "mcp[cli]>=1.0.0",
    "nats-py>=2.14.0",
    "pandas>=3.0.2",
    "protobuf>=7.34.1",
    "pyarrow>=23.0.1",
    "pydantic-ai>=0.6.2",
    "zstd>=1.5.7.3",
]

[project.scripts]
brain = "brain:main"
dummie-memory-mcp = "brain.mcp_server:mcp.run"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
